Move timezone default storage tests to a more precise package
