+++ extensions/hibernate-orm/deployment/src/test/java/io/quarkus/hibernate/orm/mapping/timezone/EntityWithTimezones.java
@@ -1,4 +1,4 @@
-package io.quarkus.hibernate.orm.mapping;
+package io.quarkus.hibernate.orm.mapping.timezone;
 
 import java.time.OffsetDateTime;
 import java.time.ZonedDateTime;
