+++ extensions/hibernate-orm/deployment/src/test/java/io/quarkus/hibernate/orm/mapping/timezone/TimezoneDefaultStorageAutoTest.java
@@ -1,8 +1,6 @@
-package io.quarkus.hibernate.orm.mapping;
+package io.quarkus.hibernate.orm.mapping.timezone;
 
 import static org.assertj.core.api.Assertions.assertThat;
-
-import java.time.ZoneId;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -11,9 +9,7 @@ import io.quarkus.hibernate.orm.SchemaUtil;
 import io.quarkus.hibernate.orm.SmokeTestUtils;
 import io.quarkus.test.QuarkusUnitTest;
 
-public class TimezoneDefaultStorageNormalizeTest extends AbstractTimezoneDefaultStorageTest {
-
-    private static final ZoneId JDBC_TIMEZONE = ZoneId.of("America/Los_Angeles");
+public class TimezoneDefaultStorageAutoTest extends AbstractTimezoneDefaultStorageTest {
 
     @RegisterExtension
     static QuarkusUnitTest TEST = new QuarkusUnitTest()
@@ -21,24 +17,23 @@ public class TimezoneDefaultStorageNormalizeTest extends AbstractTimezoneDefault
                     .addClasses(EntityWithTimezones.class)
                     .addClasses(SchemaUtil.class, SmokeTestUtils.class))
             .withConfigurationResource("application.properties")
-            .overrideConfigKey("quarkus.hibernate-orm.mapping.timezone.default-storage", "normalize")
-            .overrideConfigKey("quarkus.hibernate-orm.jdbc.timezone", JDBC_TIMEZONE.getId());
+            .overrideConfigKey("quarkus.hibernate-orm.mapping.timezone.default-storage", "auto");
 
     @Test
     public void schema() throws Exception {
         assertThat(SchemaUtil.getColumnNames(sessionFactory, EntityWithTimezones.class))
                 .doesNotContain("zonedDateTime_tz", "offsetDateTime_tz");
         assertThat(SchemaUtil.getColumnTypeName(sessionFactory, EntityWithTimezones.class, "zonedDateTime"))
-                .isEqualTo("TIMESTAMP");
+                .isEqualTo("TIMESTAMP_WITH_TIMEZONE");
         assertThat(SchemaUtil.getColumnTypeName(sessionFactory, EntityWithTimezones.class, "offsetDateTime"))
-                .isEqualTo("TIMESTAMP");
+                .isEqualTo("TIMESTAMP_WITH_TIMEZONE");
     }
 
     @Test
     public void persistAndLoad() {
         long id = persistWithValuesToTest();
-        assertLoadedValues(id, PERSISTED_ZONED_DATE_TIME.withZoneSameInstant(ZoneId.systemDefault()),
-                PERSISTED_OFFSET_DATE_TIME.withOffsetSameInstant(
-                        ZoneId.systemDefault().getRules().getOffset(PERSISTED_OFFSET_DATE_TIME.toInstant())));
+        // For some reason native storage (with H2 at least) preserves the offset, but not the zone ID.
+        assertLoadedValues(id, PERSISTED_ZONED_DATE_TIME.withZoneSameInstant(PERSISTED_ZONED_DATE_TIME.getOffset()),
+                PERSISTED_OFFSET_DATE_TIME);
     }
 }
