+++ extensions/hibernate-orm/deployment/src/test/java/io/quarkus/hibernate/orm/mapping/timezone/TimezoneDefaultStorageColumnTest.java
@@ -1,4 +1,4 @@
-package io.quarkus.hibernate.orm.mapping;
+package io.quarkus.hibernate.orm.mapping.timezone;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -9,29 +9,32 @@ import io.quarkus.hibernate.orm.SchemaUtil;
 import io.quarkus.hibernate.orm.SmokeTestUtils;
 import io.quarkus.test.QuarkusUnitTest;
 
-public class TimezoneDefaultStorageDefaultTest extends AbstractTimezoneDefaultStorageTest {
+public class TimezoneDefaultStorageColumnTest extends AbstractTimezoneDefaultStorageTest {
 
     @RegisterExtension
     static QuarkusUnitTest TEST = new QuarkusUnitTest()
             .withApplicationRoot((jar) -> jar
                     .addClasses(EntityWithTimezones.class)
                     .addClasses(SchemaUtil.class, SmokeTestUtils.class))
-            .withConfigurationResource("application.properties");
+            .withConfigurationResource("application.properties")
+            .overrideConfigKey("quarkus.hibernate-orm.mapping.timezone.default-storage", "column");
 
     @Test
     public void schema() throws Exception {
         assertThat(SchemaUtil.getColumnNames(sessionFactory, EntityWithTimezones.class))
-                .doesNotContain("zonedDateTime_tz", "offsetDateTime_tz");
+                .contains("zonedDateTime_tz", "offsetDateTime_tz")
+                // For some reason we don't get a TZ column for OffsetTime
+                .doesNotContain("offsetTime_tz");
         assertThat(SchemaUtil.getColumnTypeName(sessionFactory, EntityWithTimezones.class, "zonedDateTime"))
-                .isEqualTo("TIMESTAMP_WITH_TIMEZONE");
+                .isEqualTo("TIMESTAMP_UTC");
         assertThat(SchemaUtil.getColumnTypeName(sessionFactory, EntityWithTimezones.class, "offsetDateTime"))
-                .isEqualTo("TIMESTAMP_WITH_TIMEZONE");
+                .isEqualTo("TIMESTAMP_UTC");
     }
 
     @Test
     public void persistAndLoad() {
         long id = persistWithValuesToTest();
-        // For some reason native storage (with H2 at least) preserves the offset, but not the zone ID.
+        // For some reason column storage preserves the offset, but not the zone ID.
         assertLoadedValues(id, PERSISTED_ZONED_DATE_TIME.withZoneSameInstant(PERSISTED_ZONED_DATE_TIME.getOffset()),
                 PERSISTED_OFFSET_DATE_TIME);
     }
