+++ extensions/hibernate-orm/deployment/src/test/java/io/quarkus/hibernate/orm/mapping/timezone/TimezoneDefaultStorageNormalizeTest.java
@@ -1,6 +1,8 @@
-package io.quarkus.hibernate.orm.mapping;
+package io.quarkus.hibernate.orm.mapping.timezone;
 
 import static org.assertj.core.api.Assertions.assertThat;
+
+import java.time.ZoneId;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -9,7 +11,9 @@ import io.quarkus.hibernate.orm.SchemaUtil;
 import io.quarkus.hibernate.orm.SmokeTestUtils;
 import io.quarkus.test.QuarkusUnitTest;
 
-public class TimezoneDefaultStorageAutoTest extends AbstractTimezoneDefaultStorageTest {
+public class TimezoneDefaultStorageNormalizeTest extends AbstractTimezoneDefaultStorageTest {
+
+    private static final ZoneId JDBC_TIMEZONE = ZoneId.of("America/Los_Angeles");
 
     @RegisterExtension
     static QuarkusUnitTest TEST = new QuarkusUnitTest()
@@ -17,23 +21,24 @@ public class TimezoneDefaultStorageAutoTest extends AbstractTimezoneDefaultStora
                     .addClasses(EntityWithTimezones.class)
                     .addClasses(SchemaUtil.class, SmokeTestUtils.class))
             .withConfigurationResource("application.properties")
-            .overrideConfigKey("quarkus.hibernate-orm.mapping.timezone.default-storage", "auto");
+            .overrideConfigKey("quarkus.hibernate-orm.mapping.timezone.default-storage", "normalize")
+            .overrideConfigKey("quarkus.hibernate-orm.jdbc.timezone", JDBC_TIMEZONE.getId());
 
     @Test
     public void schema() throws Exception {
         assertThat(SchemaUtil.getColumnNames(sessionFactory, EntityWithTimezones.class))
                 .doesNotContain("zonedDateTime_tz", "offsetDateTime_tz");
         assertThat(SchemaUtil.getColumnTypeName(sessionFactory, EntityWithTimezones.class, "zonedDateTime"))
-                .isEqualTo("TIMESTAMP_WITH_TIMEZONE");
+                .isEqualTo("TIMESTAMP");
         assertThat(SchemaUtil.getColumnTypeName(sessionFactory, EntityWithTimezones.class, "offsetDateTime"))
-                .isEqualTo("TIMESTAMP_WITH_TIMEZONE");
+                .isEqualTo("TIMESTAMP");
     }
 
     @Test
     public void persistAndLoad() {
         long id = persistWithValuesToTest();
-        // For some reason native storage (with H2 at least) preserves the offset, but not the zone ID.
-        assertLoadedValues(id, PERSISTED_ZONED_DATE_TIME.withZoneSameInstant(PERSISTED_ZONED_DATE_TIME.getOffset()),
-                PERSISTED_OFFSET_DATE_TIME);
+        assertLoadedValues(id, PERSISTED_ZONED_DATE_TIME.withZoneSameInstant(ZoneId.systemDefault()),
+                PERSISTED_OFFSET_DATE_TIME.withOffsetSameInstant(
+                        ZoneId.systemDefault().getRules().getOffset(PERSISTED_OFFSET_DATE_TIME.toInstant())));
     }
 }
