+++ extensions/hibernate-orm/deployment/src/test/java/io/quarkus/hibernate/orm/mapping/timezone/TimezoneDefaultStorageNormalizeUtcTest.java
@@ -1,6 +1,8 @@
-package io.quarkus.hibernate.orm.mapping;
+package io.quarkus.hibernate.orm.mapping.timezone;
 
 import static org.assertj.core.api.Assertions.assertThat;
+
+import java.time.ZoneOffset;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -9,7 +11,7 @@ import io.quarkus.hibernate.orm.SchemaUtil;
 import io.quarkus.hibernate.orm.SmokeTestUtils;
 import io.quarkus.test.QuarkusUnitTest;
 
-public class TimezoneDefaultStorageColumnTest extends AbstractTimezoneDefaultStorageTest {
+public class TimezoneDefaultStorageNormalizeUtcTest extends AbstractTimezoneDefaultStorageTest {
 
     @RegisterExtension
     static QuarkusUnitTest TEST = new QuarkusUnitTest()
@@ -17,14 +19,12 @@ public class TimezoneDefaultStorageColumnTest extends AbstractTimezoneDefaultSto
                     .addClasses(EntityWithTimezones.class)
                     .addClasses(SchemaUtil.class, SmokeTestUtils.class))
             .withConfigurationResource("application.properties")
-            .overrideConfigKey("quarkus.hibernate-orm.mapping.timezone.default-storage", "column");
+            .overrideConfigKey("quarkus.hibernate-orm.mapping.timezone.default-storage", "normalize-utc");
 
     @Test
     public void schema() throws Exception {
         assertThat(SchemaUtil.getColumnNames(sessionFactory, EntityWithTimezones.class))
-                .contains("zonedDateTime_tz", "offsetDateTime_tz")
-                // For some reason we don't get a TZ column for OffsetTime
-                .doesNotContain("offsetTime_tz");
+                .doesNotContain("zonedDateTime_tz", "offsetDateTime_tz");
         assertThat(SchemaUtil.getColumnTypeName(sessionFactory, EntityWithTimezones.class, "zonedDateTime"))
                 .isEqualTo("TIMESTAMP_UTC");
         assertThat(SchemaUtil.getColumnTypeName(sessionFactory, EntityWithTimezones.class, "offsetDateTime"))
@@ -34,8 +34,7 @@ public class TimezoneDefaultStorageColumnTest extends AbstractTimezoneDefaultSto
     @Test
     public void persistAndLoad() {
         long id = persistWithValuesToTest();
-        // For some reason column storage preserves the offset, but not the zone ID.
-        assertLoadedValues(id, PERSISTED_ZONED_DATE_TIME.withZoneSameInstant(PERSISTED_ZONED_DATE_TIME.getOffset()),
-                PERSISTED_OFFSET_DATE_TIME);
+        assertLoadedValues(id, PERSISTED_ZONED_DATE_TIME.withZoneSameInstant(ZoneOffset.UTC),
+                PERSISTED_OFFSET_DATE_TIME.withOffsetSameInstant(ZoneOffset.UTC));
     }
 }
